upgrades to V8
