--- .babelrc
@@ -1,17 +1,0 @@
-{
-	"browserslistEnv": "javascripts",
-	"presets": [
-		[
-			"@babel/preset-env",
-			{
-				"bugfixes": true,
-				"loose": true
-			}
-		]
-	],
-	"env": {
-		"test": {
-			"browserslistEnv": "node"
-		}
-	}
-}
