--- .eslintignore
@@ -1,5 +1,0 @@
-node_modules/*
-dist/*
-bin/*
-obj/*
-Properties/*
