--- .eslintrc
@@ -1,5 +1,0 @@
-{
-    "env": {
-        "es6": true
-    }
-}
