--- .github/ISSUE_TEMPLATE/BUG_REPORT.md
@@ -1,46 +1,0 @@
----
-name: "\U0001F41B Report a bug or issue"
-about: 'Report any bugs or issues you have come across, so that we can fix it.'
----
-
-# Bug Report
-
-<!--
-Thanks for wanting to report an issue.
-
-In order for the issue to be resolved as quickly as possible please provide as detailed information as possible.
-
-Use the following headings as a guide.
--->
-
-## What is the issue?
-
-<!-- Provide a detailed description of what the issue is, including what you
-expected to happen as well as what actually happened.
--->
-
-## What steps are required to reproduce the issue?
-
-<!--
-If possible, include all steps required to recreate the issue whilst using an incognito/private browsing window.
-This will help to rule out any differences introduced via user installed extensions.
--->
-
-## What was the environment where this issue occurred?
-
-<!--
-Include the following as a minimum e.g.
-* Device: Windows Surface Pro 4
-* Operating System: Windows 10
-* Browser: Google Chrome
--->
-
-- Device:
-- Operating System:
-- Browser:
-
-## Is there anything else you think would be useful in recreating the issue?
-
-<!--
-  Screenshots, logs, repository link, supporting information etc
--->
