--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,6 +1,0 @@
-## Description
-
-## Checklist
-
-- [ ] Tested against our [testing policy](https://github.com/nhsuk/nhsuk-frontend/blob/master/docs/contributing/testing.md) (Resolution, Browser & Accessibility)
-- [ ] CHANGELOG entry
