--- .github/workflows/pull-request.yml
@@ -1,34 +1,0 @@
-name: Pull request
-
-on: pull_request
-
-jobs:
-  lint:
-    name: Code style checks
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v5
-      - uses: actions/setup-node@v6
-        with:
-          node-version-file: .nvmrc
-          cache: 'npm'
-      - name: Install dependencies
-        run: npm ci
-      - name: Run linting
-        run: npm run lint
-  tests:
-    name: Javascript tests
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        node: ['20', '22']
-    steps:
-      - uses: actions/checkout@v5
-      - uses: actions/setup-node@v6
-        with:
-          node-version: ${{ matrix.node }}
-          cache: 'npm'
-      - name: Install dependencies
-        run: npm ci
-      - name: Run javascript tests
-        run: npm run test:ci
