--- .prettierignore
@@ -1,16 +1,0 @@
-# Temporary only
-.cache/
-
-# Application
-app/
-
-# Node.js modules
-node_modules/
-
-# Build output
-public/
-
-# Files to ignore
-*.html
-*.min.js
-package-lock.json
