--- .prettierrc.json
@@ -1,23 +1,0 @@
-{
-	"arrowParens": "always",
-	"bracketSameLine": false,
-	"objectWrap": "preserve",
-	"bracketSpacing": true,
-	"semi": true,
-	"experimentalOperatorPosition": "end",
-	"experimentalTernaries": false,
-	"singleQuote": true,
-	"jsxSingleQuote": false,
-	"quoteProps": "as-needed",
-	"trailingComma": "all",
-	"singleAttributePerLine": false,
-	"htmlWhitespaceSensitivity": "css",
-	"vueIndentScriptAndStyle": false,
-	"proseWrap": "preserve",
-	"insertPragma": false,
-	"printWidth": 200,
-	"requirePragma": false,
-	"tabWidth": 2,
-	"useTabs": true,
-	"embeddedLanguageFormatting": "auto"
-}
