--- CONTRIBUTING.md
@@ -1,23 +1,0 @@
-# Contribution guidelines
-
-Thank you for your interest in contributing to NHS prototype kit, we really appreciate it. There are a number of ways that you can contribute – reporting bugs, fixing bugs, suggesting new features or writing documentation.
-
-## Contributing
-
-If you’ve got an idea or suggestion you can:
-
-- email [service-manual@nhs.net](mailto:service-manual@nhs.net)
-- get in touch on the [NHS digital service manual Slack workspace](https://nhs-service-manual.slack.com/messages/CFYL2GDGW)
-- [create a GitHub issue](https://github.com/nhsuk/nhsuk-prototype-kit/issues/new)
-
-## Raising bugs
-
-When raising bugs please explain the issue in good detail and provide a guide to how to replicate it.
-
-[Raise a bug](https://github.com/nhsuk/nhsuk-prototype-kit/issues/new?template=BUG_REPORT.md)
-
-## Suggesting features
-
-Please raise feature requests as issues before contributing any code.
-
-This ensures they are discussed properly before any time is spent on them.
