+++ app/lib/views/111/layouts/base.html
@@ -1,8 +1,9 @@
 {% set assetPath = '/nhsuk-frontend/assets' %}
 
-{% extends "template.njk" %}
+{% extends "prototype-kit-template.njk" %}
 
 {% from "attributes.njk" import nhsukAttributes %}
+
 {%- from 'action-link/macro.njk' import actionLink %}
 {%- from 'back-link/macro.njk' import backLink %}
 {%- from 'breadcrumb/macro.njk' import breadcrumb %}
@@ -39,31 +40,28 @@
 {%- from 'textarea/macro.njk' import textarea %}
 {%- from 'warning-callout/macro.njk' import warningCallout %}
 
-{% import "111/macros/prototype-kit.njk" as prototype with context %}
-{% from "111/components/callout/macro.njk" import callout with context %}
-{% from "111/components/call-to-action/macro.njk" import callToAction with context %}
-{% from "111/components/footer/macro.njk" import footer as appFooter with context %}
-{% from "111/components/card/macro.njk" import card as appCard with context %}
-{% from "111/components/card/macro.njk" import service as appService with context %}
-{% from "111/components/options-list/macro.njk" import optionsList with context %}
-{% from "111/components/sticky-bottom/macro.njk" import stickyBottom as appStickyBottom with context %}
+{%- import "111/macros/prototype-kit.njk" as prototype with context -%}
+{%- from "111/components/callout/macro.njk" import callout with context -%}
+{%- from "111/components/call-to-action/macro.njk" import callToAction with context -%}
+{%- from "111/components/footer/macro.njk" import footer as appFooter with context -%}
+{%- from "111/components/card/macro.njk" import card as appCard with context -%}
+{%- from "111/components/card/macro.njk" import service as appService with context -%}
+{%- from "111/components/options-list/macro.njk" import optionsList with context -%}
+{%- from "111/components/sticky-bottom/macro.njk" import stickyBottom as appStickyBottom with context -%}
 
-{%- from "nhsapp/styles/icon/macro.njk" import nhsappIcon %}
-{% from "nhsapp/components/bottom-navigation/macro.njk" import nhsappBottomNavigation with context %}
-{% from "nhsapp/components/global-navigation-native/macro.njk" import nhsappGlobalNavigation with context %}
+{%- from "nhsapp/styles/icon/macro.njk" import nhsappIcon -%}
+{% from "nhsapp/components/bottom-navigation/macro.njk" import nhsappBottomNavigation with context -%}
+{% from "nhsapp/components/global-navigation-native/macro.njk" import nhsappGlobalNavigation with context -%}
 
-{% set title = (page.title | default(_sublayout.title | default('title not set')) + ' &ndash; ' + serviceName) | safe %}
-{% set bodyClasses = pathclasses + (' ' if page.bodyclasses else '') + (page.bodyclasses | d('')) + (' nhs-app-native' if appframe) %}
-{% set containerClasses = page.containerClasses | default('') + (' app-width-container' if appframe) %}
-{% set mainClasses = (page.mainClasses | default(_sublayout.mainClasses | default(layout.mainClasses | default('')))) + (' app-main-wrapper' if appframe) %}
+{%- set title = (page.title | default(_sublayout.title | default('title not set')) + ' &ndash; ' + serviceName) | safe -%}
+{%- set bodyClasses = pathclasses + (' ' if page.bodyclasses else '') + (page.bodyclasses | d('')) + (' nhs-app-native' if appframe) -%}
+{%- set containerClasses = page.containerClasses | default('') + (' app-width-container' if appframe) -%}
+{%- set mainClasses = (page.mainClasses | default(_sublayout.mainClasses | default(layout.mainClasses | default('')))) + (' app-main-wrapper' if appframe) -%}
 
 {# some defalt body attributes should be added here if needed #}
-{% set bodyAttributes = (_sublayout.bodyAttributes | default({
-  "data-transition-message": "Skipping forward »",
-  "data-transition-delay": "1"
-})) | merge(page.bodyAttributes | default({})) %}
+{%- set bodyAttributes = (_sublayout.bodyAttributes | default({})) | merge(page.bodyAttributes | default({})) -%}
 
-{% set _layout = {
+{%- set _layout = {
   width: false,
   content_primary: {
     id: 'content-primary',
@@ -84,7 +82,7 @@
 {% endblock %}
 
 {# Set the page title #}
-{% block pageTitle %}{{ title }}{% endblock %}
+{%- block pageTitle %}{{ title }}{% endblock -%}
 
 {% block header %}
     {% if appframe %}
@@ -236,11 +234,8 @@
 {% endblock %}
 
 {% block bodyEnd %}
-  <script src="/nhsuk-frontend/nhsuk-frontend.min.js" type="module"></script>
-  <script type="module">
-    import { initAll } from '/nhsuk-frontend/nhsuk-frontend.min.js'
-    initAll()
-  </script>
+  {# brings through the underlying javascript from the prototype-kit-template #}
+  {{ super()}}
   {% block scripts %}
     {# Custom JavaScript files can be added to this file #}
     {% include "includes/scripts.html" %}
