+++ app/views/index.html
@@ -17,6 +17,7 @@
 {% extends page.layout.file | d(layouts.base) %}
 
 {% block content_primary %}
+
   <h1 class="nhsuk-heading-l">
     {{ page.title }}
   </h1>
