--- lib/example-templates/blank-transactional.html
@@ -1,22 +1,0 @@
-{% extends "layout.html" %}
-
-{% set mainClasses = "nhsuk-main-wrapper--s" %}
-
-{% set pageName = "Blank transactional template" %}
-
-{% block beforeContent %}
-  {{ backLink({
-    href: "#",
-    text: "Back"
-  }) }}
-{% endblock %}
-
-{% block content %}
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-two-thirds">
-      <!-- Add your content here -->
-
-
-    </div>
-  </div>
-{% endblock %}
