--- lib/example-templates/content-page.html
@@ -1,30 +1,0 @@
-{% extends "layout.html" %}
-
-{% set pageName = "Example content page" %}
-
-{% block beforeContent %}
-  {{ breadcrumb({
-    href: "#",
-    text: "Home"
-  }) }}
-{% endblock %}
-
-{% block content %}
-  <div class="nhsuk-grid-row">
-
-    <div class="nhsuk-grid-column-two-thirds">
-
-      <h1>
-        {{ pageName }}
-      </h1>
-
-      <p>
-        Visit the
-          <a href="https://service-manual.nhs.uk">NHS digital service manual</a>
-        for guidance and code examples.
-      </p>
-
-    </div>
-
-  </div>
-{% endblock %}
