--- lib/example-templates/example-templates.html
@@ -1,38 +1,0 @@
-{% extends "layout.html" %}
-
-{% set pageName = "Page templates" %}
-
-{% block beforeContent %}
-  {{ breadcrumb({
-    href: "/",
-    text: "Home"
-  }) }}
-{% endblock %}
-
-{% block content %}
-  <div class="nhsuk-grid-row">
-
-    <div class="nhsuk-grid-column-two-thirds">
-
-      <h1>{{ pageName }}</h1>
-
-      <p>You can find these templates in the <strong>/lib/example-templates</strong> folder.</p>
-
-      <p>You can copy and paste the templates in any folder within <strong>/app/views</strong>.</p>
-
-      <h2>Content pages</h2>
-
-      <ul>
-        <li><a href="/example-templates/content-page">Blank content page</a></li>
-        <li><a href="/example-templates/mini-hub">Mini hub</a></li>
-      </ul>
-
-      <h2>Transactional journeys</h2>
-
-      <ul>
-        <li><a href="/example-templates/blank-transactional">Blank transactional</a></li>
-      </ul>
-    </div>
-
-  </div>
-{% endblock %}
