--- lib/example-templates/mini-hub/index.html
@@ -1,58 +1,0 @@
-{% extends "layout.html" %}
-
-{% set pageName = "Mini hub" %}
-
-{% block beforeContent %}
-  {{ breadcrumb({
-    href: "/",
-    text: "Home"
-  }) }}
-{% endblock %}
-
-{% block content %}
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-two-thirds">
-
-      <h1>
-        Main page heading
-        <span class="nhsuk-caption-xl nhsuk-caption--bottom ">
-          Section title
-        </span>
-      </h1>
-
-      {{ contentsList({
-        items: [
-          {
-            href: "/example-templates/mini-hub/page-1",
-            text: "Main page",
-            current: "true"
-          },
-          {
-            href: "/example-templates/mini-hub/page-2",
-            text: "Page two"
-          },
-          {
-            href: "/example-templates/mini-hub/page-3",
-            text: "Page three"
-          }
-        ]
-      }) }}
-
-      <p>
-        [Page content goes here]
-      </p>
-
-      <p>
-        [Visit the
-        <a href="https://service-manual.nhs.uk">NHS digital service manual</a>
-        for guidance and examples.]
-      </p>
-
-      {{ pagination({
-        nextUrl: "/example-templates/mini-hub/page-2",
-        nextPage: "Page two"
-      }) }}
-
-    </div>
-  </div>
-{% endblock %}
