--- lib/example-templates/mini-hub/page-2.html
@@ -1,66 +1,0 @@
-{% extends "layout.html" %}
-
-{% set pageName = "Mini hub - Page two" %}
-
-{% block beforeContent %}
-  {{ breadcrumb({
-    items: [
-      {
-        href: "/",
-        text: "Home"
-      }
-    ],
-    href: "/example-templates/mini-hub",
-    text: "Main page"
-  }) }}
-{% endblock %}
-
-{% block content %}
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-two-thirds">
-
-      <h1>
-        Page two heading
-        <span class="nhsuk-caption-xl nhsuk-caption--bottom ">
-          Section title
-        </span>
-      </h1>
-
-      {{ contentsList({
-        items: [
-          {
-            href: "/example-templates/mini-hub",
-            text: "Main page"
-          },
-          {
-            href: "/example-templates/mini-hub/page-2",
-            text: "Page two",
-            current: "true"
-          },
-          {
-            href: "/example-templates/mini-hub/page-3",
-            text: "Page three"
-          }
-        ]
-      }) }}
-
-      <p>
-        [Page content goes here]
-      </p>
-
-      <p>
-        [Visit the
-        <a href="https://service-manual.nhs.uk">NHS digital service manual</a>
-        for guidance and examples.]
-      </p>
-
-      {{ pagination({
-        previousUrl: "/example-templates/mini-hub",
-        previousPage: "Main page",
-        nextUrl: "/example-templates/mini-hub/page-3",
-        nextPage: "Page two"
-      }) }}
-
-    </div>
-  </div>
-{% endblock %}
