--- lib/example-templates/question-page.html
@@ -1,53 +1,0 @@
-{% extends "layout.html" %}
-
-{% set mainClasses = "nhsuk-main-wrapper--s" %}
-
-{% set pageName = "Question goes here" %}
-
-{% block beforeContent %}
-  {{ backLink({
-    href: "#",
-    text: "Back"
-  }) }}
-{% endblock %}
-
-{% block content %}
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-two-thirds">
-
-      <form action="#" method="post" class="form">
-
-        {{ radios({
-          idPrefix: "example",
-          name: "example",
-          fieldset: {
-            legend: {
-              text: pageName,
-              classes: "nhsuk-fieldset__legend--l",
-              isPageHeading: true
-            }
-          },
-          items: [
-            {
-              value: "yes",
-              text: "Yes"
-            },
-            {
-              value: "no",
-              text: "No"
-            }
-          ]
-        }) }}
-
-        {{ button({
-          text: "Continue"
-        }) }}
-
-        <p>(See the <a href="https://service-manual.nhs.uk/design-system/components">NHS digital service manual</a> for examples.)</p>
-
-      </form>
-
-    </div>
-  </div>
-
-{% endblock %}
