--- lib/example-templates/start-page.html
@@ -1,42 +1,0 @@
-{% extends "layout.html" %}
-
-{% set pageName = serviceName %}
-
-{% block beforeContent %}
-  {{ breadcrumb({
-    href: "/",
-    text: "Home"
-  }) }}
-{% endblock %}
-
-{% block content %}
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-two-thirds">
-
-      <h1>
-        {{ pageName }}
-      </h1>
-
-      <p>Use this service to:</p>
-
-      <ul>
-        <li>do something</li>
-        <li>update your name, address or other details</li>
-        <li>do something else</li>
-      </ul>
-
-      <p>Registering takes around 5 minutes.</p>
-
-
-      <h2 class="nhsuk-heading-m">Before you start</h2>
-
-      <p>You can only use this service if you live in the UK.</p>
-
-      <a href="#" class="nhsuk-button">
-        Start now
-      </a>
-
-    </div>
-
-  </div>
-{% endblock %}
