--- lib/example_templates_routes.js
@@ -1,6 +1,0 @@
-// External dependencies
-const express = require('express');
-
-const router = express.Router();
-
-module.exports = router;
