--- lib/prototype-admin/password.html
@@ -1,54 +1,0 @@
-{% extends "layout.html" %}
-
-{% set pageName = "Enter password" %}
-{% set errors = error == "wrong-password" %}
-
-{% block content %}
-
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-two-thirds">
-      <form method="post" action="/prototype-admin/password">
-
-      {% if error == "wrong-password" %}
-        {{ errorSummary({
-          titleText: "There is a problem",
-          errorList: [
-            {
-              text: "The password is not correct",
-              href: "#password"
-            }
-          ]
-        })}}
-      {% endif %}
-
-        <h1 class="nhsuk-heading-l">
-          This is a prototype used for research
-        </h1>
-
-        <p>
-          It is not a real service. You should only continue if you have been invited to test this prototype.
-        </p>
-
-        {{ input({
-          classes: "nhsuk-input--width-10",
-          name: "password",
-          id: "password",
-          type: "password",
-          errorMessage: {
-            text: "The password is not correct"
-          } if error == "wrong-password",
-          label:{
-              text: "Password"
-          }
-        }) }}
-
-        <input type="hidden" name="returnURL" value="{{returnURL}}">
-
-        {{ button({
-          text: "Continue"
-        }) }}
-
-      </form>
-    </div>
-  </div>
-{% endblock %}
