--- lib/prototype-admin/reset-done.html
@@ -1,15 +1,0 @@
-{% extends "layout.html" %}
-
-{% set pageName = "Data reset" %}
-
-{% block content %}
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-two-thirds">
-      <h1 class="nhsuk-heading-l">{{ pageName }}</h1>
-
-      <p class="nhsuk-body">The session data has been reset.</p>
-
-      <p class="nhsuk-body"><a href="{{ returnPage }}" class="nhsuk-link">Return</a>
-    </div>
-  </div>
-{% endblock %}
