--- lib/prototype-admin/reset.html
@@ -1,27 +1,0 @@
-{% extends 'layout.html' %}
-
-{% set pageName = "Reset data" %}
-
-{% block beforeContent %}
-  {{ backLink({
-    text: "Back",
-    href: returnPage
-  }) }}
-{% endblock %}
-
-{% block content %}
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-two-thirds">
-      <h1 class="nhsuk-heading-l">{{ pageName }}</h1>
-
-      <p class="nhsuk-body">This will reset all of the data entered in this session.</p>
-
-      <form action="/prototype-admin/reset-session-data" method="post">
-        <input type="hidden" name="returnPage" value="{{ returnPage }}">
-        {{ button({
-          text: "Reset all data"
-        }) }}
-      </form>
-    </div>
-  </div>
-{% endblock %}
