--- linters/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: [
-    'airbnb-base',
-  ],
-  ignorePatterns: ['app/'],
-};
