--- owasp-dependency-check-suppressions.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<suppressions xmlns="https://jeremylong.github.io/DependencyCheck/dependency-suppression.1.3.xsd">
-    <suppress>
-        <notes><![CDATA[file name: 101:1.6.3]]></notes>
-        <packageUrl regex="true">^pkg:npm/101@.*$</packageUrl>
-        <cpe>cpe:/a:101_project:101</cpe>
-        <vulnerabilityName>CVE-2021-25943</vulnerabilityName>
-        <vulnerabilityName>GHSA-cwcx-rxgc-cmw3</vulnerabilityName>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[file name: axios:0.21.4]]></notes>
-        <packageUrl regex="true">^pkg:npm/axios@.*$</packageUrl>
-        <cpe>cpe:/a:axios:axios</cpe>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[file name: follow-redirects:1.15.1]]></notes>
-        <packageUrl regex="true">^pkg:npm/follow\-redirects@.*$</packageUrl>
-        <cpe>cpe:/a:follow-redirects:follow_redirects</cpe>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[file name: follow-redirects:1.15.1]]></notes>
-        <packageUrl regex="true">^pkg:npm/follow\-redirects@.*$</packageUrl>
-        <cpe>cpe:/a:follow-redirects_project:follow-redirects</cpe>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[file name: glob-parent:3.1.0]]></notes>
-        <packageUrl regex="true">^pkg:npm/glob\-parent@.*$</packageUrl>
-        <vulnerabilityName>CVE-2020-28469</vulnerabilityName>
-        <vulnerabilityName>GHSA-ww39-953v-wcq6</vulnerabilityName>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[file name: app.js]]></notes>
-        <packageUrl regex="true">^pkg:javascript/angularjs@.*$</packageUrl>
-        <cve>CVE-2022-25844</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[file name: semver:7.0.0]]></notes>
-        <packageUrl regex="true">^pkg:npm/semver@.*$</packageUrl>
-        <vulnerabilityName>CVE-2022-25883</vulnerabilityName>
-    </suppress>
-</suppressions>
