+++ package.json
@@ -4,19 +4,7 @@
 	"description": "Rapidly create HTML prototypes of NHS websites and services.",
 	"main": "app.js",
 	"scripts": {
-		"build": "gulp build",
-		"postinstall": "npm run build",
-		"start": "node app.js",
-		"test": "jest",
-		"test:ci": "jest --ci",
-		"prewatch": "gulp build",
-		"watch": "gulp",
-		"lint": "npm run lint:js && npm run lint:prettier",
-		"lint:fix": "npm run lint:js:fix && npm run lint:prettier:fix",
-		"lint:prettier": "prettier --cache --cache-location .cache/prettier --cache-strategy content --check .",
-		"lint:prettier:fix": "prettier --write .",
-		"lint:js": "eslint --cache --cache-location .cache/eslint --cache-strategy content --color --ignore-path .gitignore \"**/*.js\"",
-		"lint:js:fix": "npm run lint:js -- --fix"
+		"start": "node ."
 	},
 	"author": "https://github.com/nhsuk/",
 	"license": "MIT",
@@ -46,28 +34,14 @@
 		"mergician": "^2.0.2",
 		"nhsapp-frontend": "^5.0.3",
 		"nhsuk-frontend": "^10.3.1",
+		"nhsuk-prototype-kit": "^8.0.1",
 		"nunjucks": "^3.2.4",
 		"path": "^0.12.7",
 		"plugin-error": "^2.0.1",
 		"portscanner": "^2.2.0",
 		"sass-embedded": "1.77.5"
 	},
-	"devDependencies": {
-		"@typescript-eslint/eslint-plugin": "^8.46.1",
-		"@typescript-eslint/parser": "^8.43.0",
-		"eslint": "^8.57.1",
-		"eslint-config-prettier": "^10.1.8",
-		"eslint-config-standard": "^17.1.0",
-		"eslint-plugin-import": "^2.32.0",
-		"eslint-plugin-jest": "^29.0.1",
-		"eslint-plugin-jest-dom": "^5.5.0",
-		"eslint-plugin-jsdoc": "^61.1.5",
-		"eslint-plugin-n": "^16.6.2",
-		"eslint-plugin-promise": "^6.6.0",
-		"jest": "^30.1.3",
-		"jest-environment-jsdom": "^30.2.0",
-		"prettier": "^3.6.2"
-	},
+	"devDependencies": {},
 	"engines": {
 		"node": "^v24.13.0"
 	}
