--- tests/lib/core_filters.test.js
@@ -1,11 +1,0 @@
-const core_filters = require('../../lib/core_filters');
-
-test('core_filters returns correct values', () => {
-	const mockEnv = {
-		getFilter: jest.fn(),
-	};
-
-	const result = core_filters(mockEnv);
-
-	expect(typeof result.log).toBe('function');
-});
